Reorganize API and types structure for better modularity
- Create new modular structure with separated concerns:
  * `core/` - Application state, database types, result types
  * `health/` - Health check system (handlers, types, checks)
  * `api/` - Pure API layer types (response, pagination)

- Extract health system from mixed API structure:
  * Move health handlers to dedicated `health/handlers.rs`
  * Separate health check logic to `health/checks.rs`
  * Clean health-specific types in `health/types.rs`

- Improve API layer organization:
  * Split pagination logic to `api/pagination.rs`
  * Clean response types in `api/response.rs`
  * Better separation of concerns

- Maintain backward compatibility:
  * Keep all re-exports in `types.rs`
  * Update OpenAPI paths to new health module
  * All existing imports continue to work

- Benefits:
  * Better separation of concerns
  * More maintainable module structure
  * Easier to find and modify specific functionality
  * Cleaner dependency graph
  * Enhanced pagination utilities with helper methods

All tests pass, OpenAPI generation works, full quality checks pass
