+++ starter/src/api/response.rs
@@ -1,42 +1,24 @@
-// Request/Response types for API layer
-#[derive(Debug, serde::Deserialize)]
-pub struct PaginationParams {
-    pub page: Option<u32>,
-    pub limit: Option<u32>,
-}
+//! API response types and utilities
+//!
+//! This module provides standardized response types for the API layer,
+//! ensuring consistent response formats across all endpoints.
 
-impl Default for PaginationParams {
-    fn default() -> Self {
-        Self {
-            page: Some(1),
-            limit: Some(20),
-        }
-    }
-}
-
-#[derive(Debug, serde::Serialize)]
-pub struct PaginatedResponse<T> {
-    pub data: Vec<T>,
-    pub pagination: PaginationInfo,
-}
-
-#[derive(Debug, serde::Serialize)]
-pub struct PaginationInfo {
-    pub page: u32,
-    pub limit: u32,
-    pub total: u64,
-    pub total_pages: u32,
-}
-
-// API Response wrapper
+/// Standard API response wrapper
+///
+/// All successful API responses should use this structure to ensure
+/// consistency across the API surface.
 #[derive(Debug, serde::Serialize, utoipa::ToSchema)]
 pub struct ApiResponse<T> {
+    /// Whether the request was successful
     pub success: bool,
+    /// Response data (None for errors)
     pub data: Option<T>,
+    /// Optional message for additional context
     pub message: Option<String>,
 }
 
 impl<T> ApiResponse<T> {
+    /// Create a successful response with data
     pub fn success(data: T) -> Self {
         Self {
             success: true,
@@ -45,6 +27,7 @@ impl<T> ApiResponse<T> {
         }
     }
 
+    /// Create a successful response with data and message
     pub fn success_with_message(data: T, message: String) -> Self {
         Self {
             success: true,
@@ -54,14 +37,18 @@ impl<T> ApiResponse<T> {
     }
 }
 
-// Error response structure
+/// Standard error response structure
 #[derive(Debug, serde::Serialize, utoipa::ToSchema)]
 pub struct ErrorResponse {
+    /// Error details
     pub error: ErrorDetail,
 }
 
+/// Error detail information
 #[derive(Debug, serde::Serialize, utoipa::ToSchema)]
 pub struct ErrorDetail {
+    /// Error code for programmatic handling
     pub code: String,
+    /// Human-readable error message
     pub message: String,
 }
