+++ starter/src/lib.rs
@@ -2,8 +2,10 @@ pub mod api;
 pub mod auth;
 pub mod cli;
 pub mod config;
+pub mod core;
 pub mod database;
 pub mod error;
+pub mod health;
 pub mod models;
 pub mod monitoring;
 pub mod openapi;
