+++ starter/src/openapi.rs
@@ -43,11 +43,11 @@ use crate::users::models::{
     ),
     paths(
         // Health endpoints
-        crate::api::health::health,
-        crate::api::health::detailed_health,
-        crate::api::health::health_live,
-        crate::api::health::health_ready,
-        crate::api::health::health_startup,
+        crate::health::handlers::health,
+        crate::health::handlers::detailed_health,
+        crate::health::handlers::health_live,
+        crate::health::handlers::health_ready,
+        crate::health::handlers::health_startup,
 
         // Auth endpoints
         crate::auth::api::register,
